+++ apps/web/app/(app)/stats/LargestEmails.tsx
@@ -21,9 +21,10 @@ import { LargestEmailsResponse } from "@/app/api/user/stats/largest-emails/route
 import { useExpanded } from "@/app/(app)/stats/useExpanded";
 import { bytesToMegabytes } from "@/utils/size";
 import { formatShortDate } from "@/utils/date";
-import { Button } from "@/components/ui/button";
 import { getGmailUrl } from "@/utils/url";
-import { DeleteLargestEmail } from "./DeleteLargestEmail";
+import { Button, ButtonLoader } from "@/components/ui/button";
+import { onTrashMessage } from "@/utils/actions-client";
+import { useState } from "react";
 export function LargestEmails(props: { refreshInterval: number }) {
   const session = useSession();
   const { data, isLoading, error, mutate } = useSWRImmutable<
@@ -102,5 +103,42 @@ export function LargestEmails(props: { refreshInterval: number }) {
         </Card>
       )}
     </LoadingContent>
+  );
+}
+
+export function DeleteLargestEmail(props: {
+  itemId: string;
+  mutate: () => Promise<any>;
+}) {
+  const [isDeleting, setIsDeleting] = useState(false);
+  const { itemId } = props;
+  return (
+    <>
+      <Button
+        key={itemId}
+        disabled={isDeleting}
+        variant="secondary"
+        size="sm"
+        onClick={async () => {
+          if (itemId) {
+            setIsDeleting(true);
+            await onTrashMessage(itemId!);
+            await props.mutate();
+          }
+        }}
+      >
+        {isDeleting ? (
+          <>
+            <ButtonLoader />
+            Deleting...
+          </>
+        ) : (
+          <>
+            <Trash2Icon className="mr-2 h-4 w-4" />
+            Delete
+          </>
+        )}
+      </Button>
+    </>
   );
 }
